Support loading fluxd arguments from a config file
This commit makes fluxd optionally load its configuration from a given
config file, rather than relying on its command-line arguments.

This introduces the possibility of mounting config from a ConfigMap or
Secret, when running in Kubernetes; the `install` command and its
templates etc., are modified here to support this, too.

Co-authored-by: Michael Bridgen <michael@weave.works>